==> Python_Pandas/01_pandas_exercise.ipynb <==
import pandas as pd

df_raw = pd.read_excel(
                        '../product_prices.xlsx',
                        sheet_name='product_prices'
)

df_raw.head() 

df_raw_csv = pd.read_csv(
                    '../product_prices.csv',
                    sep=';',
                    decimal='.'
)

df_raw_csv.head() 

df_raw_csv.shape

df_raw_csv.shape[0]

df_raw_csv.shape[1]

df_raw_csv.head(5)

df_raw_csv.tail(10)

df_raw_csv.info()

df_raw_csv.columns

df_raw_csv['Name']

df_raw_csv[['Name', 'Value', 'Date']]

df_raw_csv.loc[:, 'Value']

df_raw_csv.loc[:, ['Value', 'Name']]

col = df_raw_csv['Value'].head() 
type(col)

col

col + 1

col * col

col_example = pd.Series(
    range(len(col))
)
col_example

col + col_example

col_example2 = pd.Series(
    range(len(col)-1)
)
col_example2 

col + col_example2

df_raw_csv['Name'].unique()

df_raw_csv['Value'].mean()

df_raw_csv['Value'].std()

df_raw_csv['Value'].median()

df_raw_csv['Value'].quantile([0, 0.1, 0.5, 0.75, 1])

df_raw_csv['Value'].describe()

columns_list = list(df_raw.columns)

columns_list

df_raw.iloc[0,:]

df_raw['Value'] = pd.to_numeric(df_raw['Value'], errors='coerce')

pd.pivot_table(df_raw, index='Name', values='Value', aggfunc='mean').sort_values(by='Value', ascending=False)
