==> Python_Pandas/01_pandas_introduction.ipynb <==
# Exercise
Read the file: `product_prices.csv` (separator ';', decimal separator '.') and do the following exercises:

1. Using the `columns` method modify the names of columns to be: `'province', 'product_types', 'currency', 'group_id', 'product_line', 'value', 'date'`<br>
1. Determine the following positional statistics: mean, standard deviation, percentiles: 0, 10%, 20%, 30%, 40%, 50%, 60%, 70%, 80%, 90%, 100%.<br>Do the exercise **in two versions**: using dedicated functions and `describe`.
1. Do you agree with this way of generating these values? Why?


import pandas as pd

# Reading the csv file
df_raw = pd.read_csv(
                    r'C:\Users\Němeček\CodersLab-Course-Python-Data-Analysis\4_Pandas\01_Data\product_prices.csv',
                    sep=';',
                    decimal='.'
    
)

# Basic information
df_raw.info() # summary table

df_raw.head(10) # data preview

# change the column names
df_raw.columns = ['province', 'product_types', 'currency', 'group_id', 'product_line', 'value', 'date']

# checking column name changes
df_raw.columns

# statistics - mean, standard deviation, percentiles: 0, 10%, 20%, 30%, 40%, 50%, 60%, 70%, 80%, 90%, 100%.
# value - mean
df_raw['value'].mean()

# value - standard deviation
df_raw['value'].std()

# value - percentiles
df_raw['value'].quantile([0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1])

# value - statistics by describe method
df_raw['value'].describe([0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1])
