==> Python_Pandas/02_Filtering data.ipynb <==
# Filtering data
While generating the `product_prices_renamed.csv` file the following processing errors occurred and they need to be found:
1. In the **date** column, data from 1888 appeared - '1888-0',
2. Check the **date** column for similar errors but with future values.
3. In the **value** column, too high value was introduced – find it and locate the row where it is (use `query`),
4. There was a spelling error in the **product_types** column for one of the products. Find it and the corresponding rows. How many such rows are there?

You do not need to assign the results to the variable: it is enough if you display them.

> Based on the solution of this task, we will later correct all errors in the data.

Hints:

Subsection 2:
1. There is only one such value.
2. Use `loc` or `query` with the condition `date > '2020-1'`.

Subsection 3:
1. There is only one such value.
1. Do the following:
a) use `describe()`, to view percentiles,
b) use `loc` or `query` to find erroneous entries,

Subsection 4:

You can do it in the following way:
a) use `unique()` method to find all available values,
b) use `loc` or `query` to find erroneous entries,
c) The number of rows can be checked with the `shape` method.

import pandas as pd

import os

# Specify the path to hte folder
folder_path = '../../01_Data/'

# List all path and directories in the folder
content = os.listdir(folder_path)

# Print the content
print(content)

data = pd.read_csv('../../01_Data/product_prices_renamed.csv', sep=';', decimal='.')

data

# 1) In the date column, data from 1888 appeared - '1888-0',

data['date'].sort_values().unique()

data[data['date'] == '1888-0']

# 2) Check the date column for similar errors but with future values.

data[data['date'] == '2099-13']

# 3) In the value column, too high value was introduced – find it and locate the row where it is (use query),

data['value'].sort_values()

data_high_value = sorted(data['value'].unique(), reverse = True)

data_high_value

data[data['value'] == 3000]

data[data['value'] == 3000].index

# 4) There was a spelling error in the product_types column for one of the products. 
# Find it and the corresponding rows. How many such rows are there?

data['product_types'].unique()

data[data['product_types'] == 'fresh chichen egges - per 666pcs.']
