==> Python_Pandas/03_Selecting data.ipynb <==
## Selection of data for analysis

Answer the following questions:

1. What was the average price of a carrot and pea mix per 1 kg in Poland in 2015?
2. What was the average price of apple juice in 2016-2018 in the Masovia province?
3. What was the average price of tomato paste in Lower Silesia Province in 2003-2015? Compared with other products does it seem reasonable to you?<br>

Suggest what can be done with values equal to 0. How does this affect the results of point 3?

> All the product groups needed for this task are in the **product_types** column.

import pandas as pd

data = pd.read_csv('../../01_Data/product_prices_renamed.csv', sep=';', decimal='.')

data

# Deleting wrong dates
data = data[~data['date'].str.contains('2099-13')]
data = data[~data['date'].str.contains('1888-0')]
data['date'] = pd.to_datetime(data['date'])

# 1) What was the average price of a carrot and pea mix per 1 kg in Poland in 2015?

data['product_types'].unique()

data['province'].unique()

data[(data['product_types'] == 'frozen carrot and pea mix - per 1kg')
    & (data['date'].dt.year == 2015)
    & (data['province'] == 'POLAND')
]['value'].mean()

# 2) What was the average price of apple juice in 2016-2018 in the Masovia province?

data[(data['product_types'] == 'apple juice, boxed - per 1l')
    & (data['date'].dt.year >= 2016)
    & (data['date'].dt.year <= 2018)
    & (data['province'] == 'MASOVIA')
]['value'].mean()

# 3) What was the average price of tomato paste in Lower Silesia Province in 2003-2015?
# Compared with other products does it seem reasonable to you?
# Suggest what can be done with values equal to 0. How does this affect the results of point 3?

data[(data['product_types'] == '30% tomato concentrate - per 1kg')
    & (data['date'].dt.year >= 2003)
    & (data['date'].dt.year <= 2015)
    & (data['province'] == 'LOWER SILESIA')
]['value'].mean()

data[(data['product_types'] == '30% tomato concentrate - per 1kg')
    & (data['date'].dt.year >= 2003)
    & (data['date'].dt.year <= 2015)
    & (data['province'] == 'LOWER SILESIA')
    & (data['value']) != 0
]['value'].mean()

df = pd.read_csv(
    '../../01_Data/product_prices_renamed.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

print(df.info())
#searching columns
print(f"Product_types:\n{df['product_types'].unique()}\n")
print(f"Provinces:\n{df['province'].unique()}")

# What was the average price of a carrot and pea mix per 1 kg in Poland in 2015?
carrot_pea_mix = df.query("product_types == 'frozen carrot and pea mix - per 1kg' and date >= '2015-1' and date <= '2015-12'")
print(f"number o rows: {carrot_pea_mix.shape[0]}")
print(f"average value: {carrot_pea_mix['value'].mean()}")
carrot_pea_mix

# What was the average price of apple juice in 2016-2018 in the Masovia province?
df_apple_juice = df.loc[
    (df['province'] == 'MASOVIA') &
    (df['product_types'] == 'apple juice, boxed - per 1l') &
    (df['date'] >= '2016-1') &
    (df['date'] <= '2018-12')
]
print(f"number o rows: {df_apple_juice.shape[0]}")
print(f"average value: {df_apple_juice['value'].mean()}")
df_apple_juice.head()

# What was the average price of tomato paste in Lower Silesia Province in 2003-2015?
# Compared with other products does it seem reasonable to you?

# with null value
df_tomato_paste = df.loc[
    (df['province'] == 'SILESIA') &
    (df['product_types'] == '30% tomato concentrate - per 1kg') &
    (df['date'] >= '2003-1') &
    (df['date'] <= '2015-12')
]
print(f"number o rows: {df_tomato_paste.shape[0]}")
print(f"average value: {df_tomato_paste['value'].mean()}")

# chechking null value
df_tomato_paste['value'].unique()

# number of null value
len(df_tomato_paste.query("value == 0"))

# without null value
df_tomato_paste = df.loc[
    (df['province'] == 'SILESIA') &
    (df['product_types'] == '30% tomato concentrate - per 1kg') &
    (df['date'] >= '2003-1') &
    (df['date'] <= '2015-12') &
    (df['value'] != 0)
]
print(f"number o rows: {df_tomato_paste.shape[0]}")
print(f"average value: {df_tomato_paste['value'].mean()}")
