==> Python_Pandas/05_Highest and lowest price.ipynb <==
# Highest and lowest price

Using `sort_values` and the data from the _product_prices_renamed.csv_ the file answer the questions below — use data aggregated for entire Poland (`province=='POLAND'`):

1. What product had the highest 'PLN' price in history? (remember to remove the value = 3000)?
1. What product achieved the lowest 'PLN' price in history? (remember to filter out values = 0),

Sort the data according to the following key:

1. by **product_types** column – descending
1. by **value** column – ascending

import pandas as pd

data = pd.read_csv('../../01_Data/product_prices_renamed.csv', sep=';', decimal='.')

data['province'].unique()

data_poland_cleaned = data[(data['province'] == 'POLAND')
    & (data['value'] != 3000)
    & (data['value'] != 0.00)
]

1) What product had the highest 'PLN' price in history? (remember to remove the value = 3000)? 

data_poland_cleaned.sort_values(
    by=['value', 'product_types'],
    ascending=[False, True],
    na_position='last'
)

2) What product achieved the lowest 'PLN' price in history? (remember to filter out values = 0),

data_poland_cleaned.sort_values(
    by=['value', 'product_types'],
    ascending=[True, True],
    na_position='last'
)

df = pd.read_csv(
  "../../01_Data/product_prices_renamed.csv",
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

df.shape[0]

df.head(5)

# What product had the highest 'PLN' price in history? (remember to remove the value = 3000)?
# dataframe without 3000 a 0
df_trimed = df.loc[
    (df['value'] != 3000) &
    (df['value'] != 0)
]
df_trimed.shape[0]

df_trimed.sort_values(
    by=['value'],
    ascending=[False], 
    na_position='last'
)

print(f"The highest value has:\n{df_trimed.loc[69229]}")

df_trimed.sort_values(
    by=['value'],
    ascending=[True], 
    na_position='last'
)

print(f"The lowest value has:\n{df_trimed.loc[910]}")

# Sort the data according to the following key:
# by product_types column – descending
# by value column – ascending

df_trimed.sort_values(
    by=['product_types', 'value'],
    ascending=[False, True], 
    na_position='last'
)
