==> Python_Pandas/06_Duplicates.ipynb <==
# Duplicates

Using the **product_prices_renamed.csv** files:

1. check how many rows are duplicated,
1. using `duplicated` check which rows were doubled.
1. using `drop_duplicates`, remove all duplicates from the output dataset (we assume that a row with a duplicate is an error and needs to be removed) - write the result to a new DataFrame.

### Hints:

#### Subsection 2:
To determine, the number of duplicate rows use the source data and the results from subsection 1.

#### Subsection 3:

1. first use `duplicated`, to find duplicated rows,
1. use `loc` to separate them from the output set,
1. use `drop_duplicates` to get only those rows that are duplicated.

import pandas as pd

data = pd.read_csv('../../01_Data/product_prices_renamed.csv', sep=';', decimal='.')

data.shape

1) check how many rows are duplicated

is_duplicated = data.duplicated()

is_duplicated[is_duplicated == True].shape

2) using duplicated check which rows were doubled.

data[is_duplicated]

3) using drop_duplicates, remove all duplicates from the output dataset (we assume that a row with a duplicate is an error and needs to be removed) - write the result to a new DataFrame.

data_without_duplicates = data.drop_duplicates(keep = 'first')

data_without_duplicates.shape
