==> Python_Pandas/07_Removing errors.ipynb <==
# Removing errors
Using the solutions from the previous class and the **product_prices_renamed.csv** file, modify `loc` so that it corrects the errors present in the dataset:

1. In the **date** column, data from 1888 appeared - '1888-0', change the value to 1999-1,
1. In the **date** column, data from 2099 appeared - '2099-13', change the value to 2019-1,
1. There is a spelling error in the **product_types** column - correct it. Number of pieces should be '10pcs.`. Check whether the task was done correctly.
1. Use `loc` to convert the values given in `EUR` to `PLN` with 4.15 exchange rate.
1. Filter from the set those rows where the price for the product is 3000.
Hint: Instead of writing `loc` twice, first query the data for rows where **currency** = `EUR` and save it to a variable.

> Remember that `loc` modifies data irrevocably.

import pandas as pd

data = pd.read_csv('../../01_Data/product_prices_renamed.csv', sep=';', decimal='.')

data

**1) In the date column, data from 1888 appeared - '1888-0', change the value to 1999-1**

data.loc[data['date'] == '1888-0', 'date'] = '1999-1'

data[data['date'] == '1999-1']

**2) In the date column, data from 2099 appeared - '2099-13', change the value to 2019-1**

data.loc[data['date'] == '2099-13', 'date'] = '2019-1'

data[data['date'] == '2019-1']

**3) There is a spelling error in the product_types column - correct it. Number of pieces should be '10pcs.`. Check whether the task was done correctly.**

data['product_types'].unique()

data.loc[data['product_types'] == 'fresh chichen egges - per 666pcs.', 'product_types'] = 'fresh chicken eggs - per 10pcs.'
data.loc[data['product_types'] == 'whole pickled cucumbers 0.9l - per 1pc.', 'product_types'] = 'whole pickled cucumbers 0.9l - per 10pcs.'

data[data['product_types'] == 'whole pickled cucumbers 0.9l - per 10pcs.'] 

**4) Use loc to convert the values given in EUR to PLN with 4.15 exchange rate.**

data.loc[data['currency'] == 'EUR', 'value'] = data.loc[data['currency'] == 'EUR', 'value'] * 4.15

data.loc[data['currency'] == 'EUR', 'currency'] = 'PLN'

data[data['currency'] == 'EUR']

**4) Filter from the set those rows where the price for the product is 3000. Hint: Instead of writing loc twice, first query the data for rows where currency = EUR and save it to a variable.**

data[data['value'] >= 3000]

# Adding a column

Following on the solution of the previous exercise, create a new column: **product** using **product_types** and **product_line**:

1. check that the **product_types** and **product_line** columns are complementary (an empty value in one column entails a non-empty value in the other).
1. create a new column: **product** using the values from the **product_types** column e.g. `df['product'] = df['product_types']`.
1. find non-empty values in the **product_line** column and enter them into the **product** column,
1. use a method of your choice to check if all values in the **product** column are non-empty.
1. remove duplicates from the table.
1. using the `to_csv` method save the data (we are going to use it later in the course), set separator=';' and `index=False`.<br>
Save the file as `product_prices_cleaned.csv`

The `read_csv` method is one of many that can be used to save a `DataFrame` as a **csv** file. Within the scope of this exercise we are interested in the following parameters.
- `sep`- row separator (default ','),
- `index`- is the index (row number by default) of the table to be saved as well (default: `True`).

Sample call:
```
df.to_csv(
    'filepath',
    sep=';', # separator setting    index=False
)
```

**1) check that the product_types and product_line columns are complementary (an empty value in one column entails a non-empty value in the other).**

data[(data['product_types'].isnull()) & (data['product_line'].isnull())]

data[(data["product_types"].notnull()) & (data["product_line"].notnull())]

**2) create a new column: product using the values from the product_types column e.g. df['product'] = df['product_types'].**

data['product'] = data['product_types']

**3) find non-empty values in the product_line column and enter them into the product column,**

data.loc[data['product_types'].isnull(), 'product'] = data['product_line']

**4) use a method of your choice to check if all values in the product column are non-empty.**

data[data['product'].isnull()]

**5) remove duplicates from the table.**

data.shape

data = data.drop_duplicates()

data.shape

**6) using the to_csv method save the data (we are going to use it later in the course), set separator=';' and index=False.
Save the file as product_prices_cleaned.cs**v

data.to_csv(
    '../../01_Data/product_prices_cleaned.csv', # path to file
    sep=';', # separator setting
    index=False
)
