==> Python_Pandas/08_Adding a column.ipynb <==
# Adding a column

Following on the solution of the previous exercise, create a new column: **product** using **product_types** and **product_line**:

1. check that the **product_types** and **product_line** columns are complementary (an empty value in one column entails a non-empty value in the other).
1. create a new column: **product** using the values from the **product_types** column e.g. `df['product'] = df['product_types']`.
1. find non-empty values in the **product_line** column and enter them into the **product** column,
1. use a method of your choice to check if all values in the **product** column are non-empty.
1. remove duplicates from the table.
1. using the `to_csv` method save the data (we are going to use it later in the course), set separator=';' and `index=False`.<br>
Save the file as `product_prices_cleaned.csv`

The `read_csv` method is one of many that can be used to save a `DataFrame` as a **csv** file. Within the scope of this exercise we are interested in the following parameters.
- `sep`- row separator (default ','),
- `index`- is the index (row number by default) of the table to be saved as well (default: `True`).

Sample call:
```
df.to_csv(
    'filepath',
    sep=';', # separator setting    index=False
)
```

import pandas as pd

df = pd.read_csv(
  '../../01_Data/product_prices_renamed.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

df.loc[df["date"] == "1888-0", "date"] = "1999-1"
df.loc[df["date"] == "2099-13", "date"] = "2019-1"

df.loc[df["currency"] == "EUR", "value"] = df["value"] * 4.15

df.loc[df["currency"] == "EUR", "currency"] = "PLN"

df = df[~(df["value"] == 3000)]

df

complementary = ((df['product_types'].isnull() & df['product_line'].notnull()) |
                 (df['product_types'].notnull() & df['product_line'].isnull()))

df['product'] = df['product_types']

df['product'].fillna(df['product_line'], inplace=True)

df

if df['product'].isnull().any():
    print("There are empty values in the 'product' column")
else:
    print("All values in the 'product' column are non-empty.")

df.drop_duplicates(inplace=True)

df.to_csv('product_prices_cleaned.csv', sep=';', index=False)

df
