==> Python_Pandas/09_Grouping products.ipynb <==
# Grouping products

Using the data from the  **product_prices_cleaned.csv** file for provinces, do the following:

1. What was the average monthly price of each commodity?
1. Which product had the highest price volatility over the years?

Use the **product** and **value** columns for analysis.

Additionally:

1. Consider whether any more assumptions are needed for the tasks.
1. Why can this task be done only now, after cleaning the data? Do you think there are any more operations needed?

import pandas as pd

data = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

data_by_date = data.groupby('date')

data_by_date

for (key, data_date) in data_by_date:
    print(key)
    print(data_date)

**Display the available groups:**

data_by_date.groups.keys()

**get_group() - selecting the subset we are interested in.**

data_by_date.get_group('1999-1')

data

data_by_product = data.groupby(by=["product","date"])

data_by_product.groups.keys()

data_by_product.get_group(('30% tomato concentrate - per 1kg', '1999-1'))['value'].mean()

data_by_product.get_group(("Backpacker's canned pork meat - per 300 g", '2008-4'))['value'].median()

for (key, data_by) in data_by_product:
    print(key, data_by['value'].mean()) 

data_by_product[['value']].agg(['min', 'max', 'mean'])

max_price_per_product = data_by_product['value'].max()

max_price_product = max_price_per_product.idxmax()

max_price_product

max_price_value = max_price_per_product.max()

max_price_value

print(f"Highest price: {max_price_value} - product: {max_price_product}")

price_volatility = data_by_product['value'].std()

volatility_by_product = data.groupby('product')['value'].std()

most_volatile_product = volatility_by_product.idxmax()
highest_volatility = volatility_by_product.max()

print(f"Highest volatility: {highest_volatility} -  product: {most_volatile_product}")
