==> Python_Pandas/10_Aggregation.ipynb <==
# Aggregation


Using data from file **product_prices_cleaned.csv**, aggregate the data for each product by month and determine the statistics: `min, max, median, mean, std` for prices (**value** column):

1. skip the national data in the analysis,
1. directly on the object from `groupby`,
1. write a loop that will calculate these values for individual provinces.

Use the `agg` method, and aggregate the data with the `'product','date'` column to complete the exercise.

import pandas as pd

data = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

data['date'] = pd.to_datetime(data['date'])

data = data[data['province'] != 'NATIONAL']

grouped = data.groupby(
    [
        'product',
        'province',
        pd.Grouper(key='date', freq='M')
    ]
)

aggregated_data = grouped['value'].agg(
    min='min',
    max='max',
    median='median',
    mean='mean',
    std='std'
).reset_index()

aggregated_data

aggregated_data.to_csv("aggregated_product_prices.csv", index=False)
