==> Python_Pandas/11_Dataset normalization.ipynb <==
# Dataset normalization

In the sheets of the **dims.xlsx** file there are dictionaries for the data from the **product_prices_cleaned.csv** file. Use `merge` to normalize the data following the steps:

1. Read the contents of the **dims.xlsx** file sheets to separate `DataFrames`.
For readability base names of frames on the names of sheets.

1. Read the data from **product_prices_cleaned.csv** file to the `df` variable.

1. Based on the **d_province** workbook, use the `id` column to add the `province_id` column to the `df` frame.

1. Based on the  **d_product** workbook, add the `product_id` column to the `df` frame.

1. From the table, extract only the columns that refer to other tables, e.g.. **product_id** and the columns **value**, **date**. Do you think this is more readable? What are potential benefits of this approach?

> We will tell you how to read many workbooks at once when we discuss `openpyxl`.

You can find more about database normalization at the [link](https://www.sqlshack.com/what-is-database-normalization-in-sql-server/).

import pandas as pd

data_dims_product = pd.read_excel(
    r'../../01_Data\dims.xlsx',
    sheet_name = 'd_product'
)

data_dims_province = pd.read_excel(
    r'../../01_Data\dims.xlsx',
    sheet_name = 'd_province'
)

data_dims_product_group = pd.read_excel(
    r'../../01_Data\dims.xlsx',
    sheet_name = 'd_product_group'
)

data_cleaned = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

join_province = pd.merge(
    left = data_cleaned,
    right = data_dims_province,
    how='left',
    left_on=['province'],
    right_on=['province']
)

join_province

data_dims_product

data_dims_product.rename(columns={'product': 'product_line'}, inplace=True)

data_dims_product

join_province_product = pd.merge(
    left = join_province,
    right = data_dims_product,
    how='left',
    left_on=['product_group_id'],
    right_on=['product_group_id']
)

join_province_product

# I can only get rid of duplicate columns through column selection
join_province_product[['province', 'product_types', 'currency', 'product_group_id', 'product_line_x', 'value', 'date', 'province_id', 'product_id']]

# From the table, extract only the columns that refer to other tables, e.g.. product_id and the columns value, date. 
join_province_product[['product_id', 'currency', 'product_group_id', 'value', 'date', 'province_id']]
