==> Python_Pandas/12_Merge with conditions.ipynb <==
# Merge with conditions

Using the raw data from the **product_prices_cleaned.csv** file, perform the task of finding out how many times a lower price has been quoted for a given product, province, month historically. To do this, perform the following steps:

1. merge the table with itself. What type of join should be used?
1. filter the data to find earlier years and the values smaller than the one from current year in a given province.
1. group the data accordingly.

Which product(s) had the most such occurrences?

> See what happens to the column names when the columns are not used as a merge condition, but have the same names

import pandas as pd

data = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

data['date'] = pd.to_datetime(data['date'])

# Merge the table with itself
merged = pd.merge(
    data,
    data,
    on=["product", "province"],
    suffixes=("_x", "_y")
)

# Filter to find historical years and lower prices
filtered = merged[
    (merged['date_x'] > merged['date_y']) &  # Historical year
    (merged['value_y'] < merged['value_x'])  # Lower price
]

filtered.head(5)

# Group by product and province and count occurrences
result = (
    filtered
    .groupby(['product', 'province'])
    .size()
    .reset_index(name='lower_price_count')
)

result

most_occurrences = result.sort_values(by='lower_price_count', ascending=False)

most_occurrences

print("Products with the most occurrences of lower historical prices:")
print(most_occurrences.head())
