==> Python_Pandas/13_formatting dates.ipynb <==
# Formatting dates

Load the _product_prices_cleaned.csv_ file, the _date_ column contains information about the months of value reporting.

1. What date format is it?
1. Change the column format to `datetime64`.
1. Add a new column – _month_ to the DataFrame as a way to isolate the information about the month from the _date_ column.
1. Add a new column – _quarter_ to the DataFrame as a way to isolate the information about the quarter from the _date_ column,
1. Add a new column – _year_ to the DataFrame as a way to isolate the information about the year from the _date_ column,
1. Using the `dt.strfime` method ([link](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.dt.strftime.html) to the documentation), convert the  _date_ column to the _YYYY-MM-01_ format and overwrite its value; e.g.:
```
df['date'] = col.dt.strftime('format')
```
1. Overwrite the _product_prices_cleaned.csv_ file. It will be used in the following sections to analyze the collection.


import pandas as pd

dates = ['2020-02', '2013-11', '2029-06']
extraction_example = pd.Series(pd.to_datetime(dates, format='%Y-%m'))
extraction_example.dt.month  # extracting months
extraction_example.dt.year  # extracting only years

data = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=';',
  encoding='UTF-8',
  decimal='.'
)

data.head()

data['date'].head()

data['date'] = pd.to_datetime(data['date'], format='%Y-%m')

data.head()

"""Add a new column – month to the DataFrame as a way to isolate the information about the month from the date column.
Add a new column – quarter to the DataFrame as a way to isolate the information about the quarter from the date column,
Add a new column – year to the DataFrame as a way to isolate the information about the year from the date column,"""

data_year = data['date'].dt.year

data_year.head()

data['year'] = data_year

data_month = data['date'].dt.month

data['month'] = data_month

data.head()

data_quarter = data['date'].dt.quarter

data_quarter.head()

data['quarter'] = data_quarter

data.head()

data['date'] = data['date'].dt.strftime('%Y-%m-%d')

data

data.to_csv('../../01_Data/product_prices_cleaned.csv', index=False)
