==> Python_Pandas/15_openpxl_File conversion.ipynb <==
# File conversion

Using `openpyxl` convert the **product_prices_cleaned.csv** file to an Excel file.

import pandas as pd
from openpyxl import Workbook

data = pd.read_csv(
    '../../01_Data/product_prices_cleaned.csv',
    sep=',',
    encoding='UTF-8',
    decimal='.'
)

data.head(5)

workbook = Workbook()
sheet = workbook.active
sheet.title = "Product Prices"

# Write the column headers
for col_idx, column in enumerate(data.columns, start=1):
    sheet.cell(row=1, column=col_idx, value=column)

# Write the data to the Excel sheet
for row_idx, row in enumerate(data.itertuples(index=False), start=2):
    for col_idx, value in enumerate(row, start=1):
        sheet.cell(row=row_idx, column=col_idx, value=value)


# Save the workbook as an Excel file
excel_file = "product_prices_cleaned.xlsx"
workbook.save(excel_file)

print(f"File has been successfully converted to {excel_file}")
