==> Python_Pandas/16_openpxl_Grouping data across sheets.ipynb <==
# Grouping data across sheets

Modify the previous exercise so that each product group is in a separate worksheet together with the corresponding `product_group_id`. For example
`product_group_id` 1 should go to the sheet named `1`.

Use the method to create a new workbook:
```
wb.create_sheet(name, index)
```
Where:
- `name` - is the name of the workbook, where we want to save the data (should be a string),
- `index` - is the worksheet position in the workbook.

import pandas as pd
from openpyxl import Workbook

data = pd.read_csv(
  '../../01_Data/product_prices_cleaned.csv',
  sep=',',
  encoding='UTF-8',
  decimal='.'
)

data.head()

grouped_data = data.groupby("product_group_id")

# Create a new Excel workbook
wb = Workbook()

# Remove the default sheet created by openpyxl
default_sheet = wb.active
wb.remove(default_sheet)

# Add each product group to a separate sheet
for group_id, group_data in grouped_data:
    sheet_name = str(group_id)
    sheet = wb.create_sheet(title=sheet_name)
    
    for col_idx, column in enumerate(group_data.columns, start=1):
        sheet.cell(row=1, column=col_idx, value=column)
    
    for row_idx, row in enumerate(group_data.itertuples(index=False), start=2):
        for col_idx, value in enumerate(row, start=1):
            sheet.cell(row=row_idx, column=col_idx, value=value)

# Save the workbook
excel_file = "grouped_product_data.xlsx"
wb.save(excel_file)
