==> Python_Pandas/18_pandas_Analytical _functions_without_aggregation.ipynb <==
import pandas as pd

df = pd.read_csv(
    './01_Data/product_prices_cleaned.csv',
    sep=','
)
df.head()

df_test = df.loc[
    (df['product'] == 'barley groats sausage - per 1kg') &
    (df['province'] == 'POLAND')
]
df_test.head() 

df_sorted = df_test.sort_values(by=['value'], ascending=False)

df_sorted['rn'] = df_sorted.groupby(by=['product']).cumcount()

df_sorted.head()

df_sorted.query("rn < 10")

df_sorted['cs'] = df_sorted.groupby(by=['product'])['value'].cumsum()
df_sorted.head()

df_sorted = df_sorted.sort_values(by=['date'], ascending=True)

df_sorted['cummin'] = df_sorted.groupby(by=['product'])['value'].cummin()
df_sorted.head()

df_sorted['cummax'] = df_sorted.groupby(by=['product'])['value'].cummax()

df_sorted.head()
